Fixed the regressions caused by the new rendering mechanism
In 1.9, the course format plugin was executed before the block content
was populated. So we could read the current course display directly from
the database. In 2.x, the block content is populated before the course
format is executed so we must must do a bit of dark magic and try to
guess what the course format will do according the passed HTTP params.

For the record, I know it is a bloody hack. At the moment, I am not able
to come with a better solution though ...