Added global settings to plugin
This is a fix to issues #16 and #17.
I added a setting possibility to decide if the titles should be forced
to use or not (then the summary would be taken if the default title is
set and the summary is not empty). This should fix issue #17
And then I added a settings.php to enable global settings. There the
admin can set the default value and is able to lock the settings so
that teachers cannot change value anymore. This setting simply will not
be presented to the teacher in the individual block settings. This
should fix and extent issue #16.